MCPEvents: support variable framerates (#17)
* Handle variable framerates

The previous implementation assumed framerates were consistent,
which isn't the case if we drop frames due to RTSP dropout or similar.

When this happens, we don't get a framerate fps value from the
video reader.

To calculate the output FPS for the written video, calculate by
reading every video frame and corresponding timestamp.

Then for frame annotation, use the caps value returned for the
time offset of each frame.

One other minor change in this commit to get rid of warnings for
invalid codec when using caps version of MP4V string instead of
lowercase.  This likely wasn't related but is probably the right
change to make

* Handle multiple events per frame

When frames are dropped in the video this looks like multiple
events which occur on a shorter range of frames.

When this happens, place the annotation on whatever frame happens
before the event in the case that the next frame is after the event.

A better way might be to try to find the closest frame in
timestamp however the approach proposed here is simpler and
works when the total time gap of dropped frames is relatively small.